separate webview tests for from commands tests
